test Spring read JAR
