+++ chartjs/package.json
@@ -20,6 +20,7 @@
     "@angular/platform-browser": "^6.0.0",
     "@angular/platform-browser-dynamic": "^6.0.0",
     "@angular/router": "^6.0.0",
+    "angular-core": "0.0.5",
     "chart.js": "^2.7.2",
     "core-js": "^2.5.4",
     "rxjs": "^6.0.0",
