--- src/data/problem/temp/CreateInputItems.java
@@ -1,26 +1,0 @@
-package data.problem.temp;
-
-import java.util.List;
-
-public class CreateInputItems {
-	private List<InputsNameTable> inputsNameTable;
-
-	public CreateInputItems() {
-		super();
-	}
-
-	public CreateInputItems(List<InputsNameTable> inputsNameTable) {
-		super();
-		this.inputsNameTable = inputsNameTable;
-	}
-
-	public List<InputsNameTable> getInputsNameTable() {
-		return inputsNameTable;
-	}
-
-	public void setInputsNameTable(List<InputsNameTable> inputsNameTable) {
-		this.inputsNameTable = inputsNameTable;
-	}
-	
-	
-}
