--- src/data/problem/temp/Faqs.java
@@ -1,5 +1,0 @@
-package data.problem.temp;
-
-public class Faqs {
-
-}
