--- src/data/problem/temp/Introduction.java
@@ -1,5 +1,0 @@
-package data.problem.temp;
-
-public class Introduction {
-
-}
