--- src/data/problem/temp/Restrictions.java
@@ -1,26 +1,0 @@
-package data.problem.temp;
-
-import java.util.List;
-
-public class Restrictions {
-	private List<RestrictionsList> restrictionsList;
-
-	public Restrictions(List<RestrictionsList> restrictionsList) {
-		super();
-		this.restrictionsList = restrictionsList;
-	}
-	
-	public Restrictions() {
-		
-	}
-
-	public List<RestrictionsList> getRestrictionsList() {
-		return restrictionsList;
-	}
-
-	public void setRestrictionsList(List<RestrictionsList> restrictionsList) {
-		this.restrictionsList = restrictionsList;
-	}
-
-	
-}
