+++ src/data/submission/CreateInputItems.java
@@ -1,0 +1,26 @@
+package data.submission;
+
+import java.util.List;
+
+public class CreateInputItems {
+	private List<InputsNameTable> inputsNameTable;
+
+	public CreateInputItems() {
+		super();
+	}
+
+	public CreateInputItems(List<InputsNameTable> inputsNameTable) {
+		super();
+		this.inputsNameTable = inputsNameTable;
+	}
+
+	public List<InputsNameTable> getInputsNameTable() {
+		return inputsNameTable;
+	}
+
+	public void setInputsNameTable(List<InputsNameTable> inputsNameTable) {
+		this.inputsNameTable = inputsNameTable;
+	}
+	
+	
+}
