--- src/data/submission/Faq.java
@@ -1,9 +1,0 @@
-package data.submission;
-
-/**
- * @author skner
- *
- */
-public class Faq {
-
-}
