+++ src/data/submission/Faqs.java
@@ -1,0 +1,5 @@
+package data.submission;
+
+public class Faqs {
+
+}
