+++ src/data/submission/Introduction.java
@@ -1,9 +1,5 @@
 package data.submission;
 
-/**
- * @author skner
- *
- */
 public class Introduction {
 
 }
